requirements.txt
sentence-transformers
faiss-cpu
numpy
pickle-mixin